Cast to int16 afterwards to maintain the same data type as input

+++ calibrate_inl.py
@@ -32,7 +32,7 @@ def main(args):
             data_corrected = uncal_data / correction
             
             # Save the corrected data to a new FITS file
-            hdu = fits.PrimaryHDU(data_corrected)
+            hdu = fits.PrimaryHDU(data_corrected.astype(np.uint16))
             hdu.writeto(os.path.join(output_path, f'{fits_file}'), overwrite=True)
